SFTP protocol proposal for publish only.
This is just a proposal to see how much of a lift the addition of a new
protocol costs in terms of effort from a dev not entirely familiar with the
conventions, we'll still need to run it by Dan and Ram to see if this is
actually going in the right direction. I have a feeling it's maybe leaking some
abstractions they put in place?